feat/#11: Restructure packages
